Remove restriction on assigning Admin role during registration
The code changes in `AccountController.cs` remove the validation that prevents users from being assigned the Administrator role (RoleId = 1) during registration. Specifically, the lines that check if `registerUserDto.RoleId` is equal to 1 and return a `BadRequest` response if true have been deleted. This means that the restriction on assigning the Administrator role to users during registration has been lifted.