Add BadRequestException class to handle bad request errors
A new custom exception class `BadRequestException` has been created within the `PersonalWebApi.Exceptions` namespace. This class extends the base `Exception` class and includes a constructor that accepts a message string, which is then passed to the base class constructor. This custom exception can be used to handle bad request errors in the PersonalWebApi project.