Remove rolesAccessConfig.json from project configuration
The changes in the `PersonalWebApi.csproj` file involve the removal of specific configurations related to the `rolesAccessConfig.json` file. Specifically, the lines that remove the `rolesAccessConfig.json` file from the content and compile items have been deleted. This means that the `rolesAccessConfig.json` file is no longer being explicitly managed as a content or compile item in the project file.