Add "Template" section to README.md
A new "Template" section has been added to the README.md file. This section informs users about the `TemplateAPI-branch`, which serves as a template with the main API features. It includes instructions on how to use this template for a new API project, such as changing the project name, solution name, namespaces, and reading the 'Getting Started' section from the branch's README.md.