Add CreateItemAsync method to CosmosDbContentStoreService
A new method `CreateItemAsync` has been added to the
`CosmosDbContentStoreService` class in the
`PersonalWebApi.Services.Azure` namespace. This method takes an
`object` as a parameter and checks if it is null. If it is null, it
throws an `ArgumentException`. If it is not null, it calls the
`_container.CreateItemAsync` method to create the item in the Cosmos
DB container.

This new method simplifies the item creation process when the
partition key is not necessary.