Add ChatHistoryShortTermDeleteDocumentDto class
A new class `ChatHistoryShortTermDeleteDocumentDto` has been added to the `PersonalWebApi.Models.Models.Memory` namespace. This class inherits from `ChatHistoryShortTermMessageDto` and includes a `FileId` property of type `string`, decorated with the `JsonProperty` attribute to specify the JSON property name as "fileId". A constructor is defined to initialize the class with `conversationUuid` and `sessionUuid` parameters. The `using Newtonsoft.Json;` directive has been added to the file.