Add CheckConversationAccessFilter for access control
A new action filter class `CheckConversationAccessFilter` has been added to the `PersonalWebApi.ActionFilters` namespace. This filter implements the `IActionFilter` interface, providing logic to execute before and after an action method is called.

In the `OnActionExecuting` method:
- The `conversationUuid` is extracted from the action arguments and injected into `HttpContext.Items`.
- A new `sessionId` is generated and added to `HttpContext.Items`.
- The user's access to the conversation is checked using the `CheckUserAccessToConversation` method.
- If the user does not have access, the result is set to a 403 Forbidden response.

The `OnActionExecuted` method is included but currently does not perform any actions.

The `CheckUserAccessToConversation` method contains placeholder logic to determine if the user has access to the conversation based on the `conversationUuid` and the user's authentication status.