Add UploadFileFromUriToFileStorageRequestDto class
A new class `UploadFileFromUriToFileStorageRequestDto` has been added to the `PersonalWebApi.Models.Storage` namespace. This class is designed for uploading a file from a URL to an Azure Blob Storage Account. It includes the following properties:

- `FileUri` (string): The URI of the file to be uploaded. This property is required.
- `FileName` (string): The name of the file to be uploaded. This property is required.
- `Overwrite` (bool): A flag indicating whether to overwrite the file if it already exists. The default value is `true`.
- `Metadata` (Dictionary<string, string>?): Optional metadata for the file. Note that metadata does not work with Swagger and is always null.

The `System.ComponentModel.DataAnnotations` namespace is used to apply the `[Required]` attribute to the `FileUri` and `FileName` properties.