Add UploadFileToFileStorageRequestDto class
Added a new using directive for System.ComponentModel.DataAnnotations.
Introduced a new namespace PersonalWebApi.Models.Storage.
Created a new class UploadFileToFileStorageRequestDto within the
PersonalWebApi.Models.Storage namespace.
Added a summary comment to describe the purpose of the
UploadFileToFileStorageRequestDto class, specifically for uploading
files to an Azure Blob Storage Account and noting that metadata does
not work with Swagger.
Added a File property of type IFormFile with a [Required] attribute to
ensure it is mandatory.
Added an Overwrite property of type bool with a default value of true.
Added a Metadata property of type Dictionary<string, string>? which is
nullable.