Add SpecifyDocumentLanguageStep class with async method
Introduce SpecifyDocumentLanguageStep class in the
PersonalWebApi.Processes.Document.Steps namespace, marked as
experimental with [Experimental("SKEXP0080")]. This class inherits
from KernelProcessStep and includes the SpecifyDocumentLanguageAsync
method, decorated with [KernelFunction]. The method determines the
document language using DocumentInfoPlugin and updates
documentStepDto with the language and an event. Necessary using
directives are added.